refactor(galgame): extract ocr window scanner from ocr_reader
